remove useless ExtractorUtils
