--- shardingsphere-sql-parser/shardingsphere-sql-parser-engine/src/test/java/org/apache/shardingsphere/sql/parser/core/extractor/util/ExtractorUtilsTest.java
@@ -1,39 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.shardingsphere.sql.parser.core.extractor.util;
-
-import org.antlr.v4.runtime.ParserRuleContext;
-import org.junit.Test;
-
-import java.util.Collection;
-
-import static org.hamcrest.CoreMatchers.is;
-import static org.junit.Assert.assertFalse;
-import static org.junit.Assert.assertThat;
-
-public final class ExtractorUtilsTest {
-    
-    @Test
-    public void assertGetAllDescendantNodesUsingNullParserRuleContextChild() {
-        ParserRuleContext parserRuleContext = new ParserRuleContext();
-        parserRuleContext.addAnyChild(null);
-        Collection<ParserRuleContext> collection = ExtractorUtils.getAllDescendantNodes(parserRuleContext, RuleName.ADD_COLUMN_SPECIFICATION);
-        assertThat(collection.size(), is(0));
-        assertFalse(collection.contains(parserRuleContext));
-    }
-}
